* split CPS1 scanning and instrset/samplecoll logic into their own files * CPSScanner -> CPS2Scanner, even if it encompasses CPS3 and some ZN-1/ZN-2 games
