+++ src/main/formats/CPS2Format.h
@@ -1,7 +1,8 @@
 #pragma once
 #include "Format.h"
 #include "VGMColl.h"
-#include "CPSScanner.h"
+#include "CPS2Scanner.h"
+#include "CPS1Scanner.h"
 
 enum CPSSynth {
   QSOUND,
@@ -11,6 +12,10 @@ enum CPSSynth {
 
 constexpr int CPS1_OKIMSM6295_SAMPLE_RATE = 7576; // 1Mhz / 132, rounded
 
-BEGIN_FORMAT(CPS)
-  USING_SCANNER(CPSScanner)
+BEGIN_FORMAT(CPS1)
+  USING_SCANNER(CPS1Scanner)
+END_FORMAT()
+
+BEGIN_FORMAT(CPS2)
+  USING_SCANNER(CPS2Scanner)
 END_FORMAT()
