+++ src/main/formats/CPS2Scanner.cpp
@@ -1,33 +1,53 @@
 #include "pch.h"
 #include "Root.h"
-#include "CPSScanner.h"
+#include "CPS2Scanner.h"
 #include "CPSSeq.h"
-#include "CPSInstr.h"
+#include "CPS2Instr.h"
 #include "MAMELoader.h"
+#include "CPS2Format.h"
 
 using namespace std;
 
-void CPSScanner::Scan(RawFile *file, void *info) {
+CPSFormatVer GetVersionEnum(string &versionStr) {
+  if (versionStr == "CPS1_2.00") return VER_CPS1_200;
+  if (versionStr == "CPS1_2.00ff") return VER_CPS1_200ff;
+  if (versionStr == "CPS1_3.50") return VER_CPS1_350;
+  if (versionStr == "CPS1_4.25") return VER_CPS1_425;
+  if (versionStr == "CPS1_5.00") return VER_CPS1_500;
+  if (versionStr == "CPS1_5.02") return VER_CPS1_502;
+  if (versionStr == "1.00") return VER_100;
+  if (versionStr == "1.01") return VER_101;
+  if (versionStr == "1.03") return VER_103;
+  if (versionStr == "1.04") return VER_104;
+  if (versionStr == "1.05a") return VER_105A;
+  if (versionStr == "1.05c") return VER_105C;
+  if (versionStr == "1.05") return VER_105;
+  if (versionStr == "1.06b") return VER_106B;
+  if (versionStr == "1.15c") return VER_115C;
+  if (versionStr == "1.15") return VER_115;
+  if (versionStr == "1.16b") return VER_116B;
+  if (versionStr == "1.16") return VER_116;
+  if (versionStr == "1.30") return VER_130;
+  if (versionStr == "1.31") return VER_131;
+  if (versionStr == "1.40") return VER_140;
+  if (versionStr == "1.71") return VER_171;
+  if (versionStr == "1.80") return VER_180;
+  if (versionStr == "2.00") return VER_200;
+  if (versionStr == "2.01b") return VER_201B;
+  if (versionStr == "2.10") return VER_210;
+  if (versionStr == "2.11") return VER_211;
+  if (versionStr == "CPS3") return VER_CPS3;
+  return VER_UNDEFINED;
+}
+
+void CPS2Scanner::Scan(RawFile *file, void *info) {
   MAMEGame *gameentry = (MAMEGame *) info;
   CPSFormatVer fmt_ver = GetVersionEnum(gameentry->fmt_version_str);
 
   if (fmt_ver == VER_UNDEFINED) {
     string alert = "XML entry uses an undefined QSound version: " + gameentry->fmt_version_str;
-    pRoot->AddLogItem(new LogItem(alert, LOG_LEVEL_ERR, "CPSScanner"));
+    pRoot->AddLogItem(new LogItem(alert, LOG_LEVEL_ERR, "CPS2Scanner"));
     return;
-  }
-
-  switch (fmt_ver) {
-    case VER_CPS1_200:
-    case VER_CPS1_200ff:
-    case VER_CPS1_350:
-    case VER_CPS1_425:
-    case VER_CPS1_500:
-    case VER_CPS1_502:
-      LoadCPS1(gameentry);
-      return;
-    default:
-      break;
   }
 
   MAMERomGroup *seqRomGroupEntry = gameentry->GetRomGroupOfType("audiocpu");
@@ -85,8 +105,8 @@ void CPSScanner::Scan(RawFile *file, void *info) {
       break;
   }
 
-  CPSInstrSet *instrset = 0;
-  CPSSampColl *sampcoll = 0;
+  CPS2InstrSet *instrset = 0;
+  CPS2SampColl *sampcoll = 0;
   CPSSampleInfoTable *sampInfoTable = 0;
   CPSArticTable *articTable = 0;
 
@@ -136,7 +156,7 @@ void CPSScanner::Scan(RawFile *file, void *info) {
     }
   }
 
-  instrset = new CPSInstrSet(programFile,
+  instrset = new CPS2InstrSet(programFile,
                              fmt_ver,
                              instr_table_offset,
                              num_instr_banks,
@@ -147,7 +167,7 @@ void CPSScanner::Scan(RawFile *file, void *info) {
     delete instrset;
     instrset = NULL;
   }
-  sampcoll = new CPSSampColl(samplesFile, instrset, sampInfoTable, 0, 0, sampcoll_name);
+  sampcoll = new CPS2SampColl(samplesFile, instrset, sampInfoTable, 0, 0, sampcoll_name);
   if (!sampcoll->LoadVGMFile()) {
     delete sampcoll;
     sampcoll = NULL;
@@ -169,7 +189,7 @@ void CPSScanner::Scan(RawFile *file, void *info) {
   }
 
   // Add SeqTable as Miscfile
-  VGMMiscFile *seqTable = new VGMMiscFile(CPSFormat::name, seqRomGroupEntry->file, seq_table_offset, seq_table_length, seq_table_name);
+  VGMMiscFile *seqTable = new VGMMiscFile(CPS2Format::name, seqRomGroupEntry->file, seq_table_offset, seq_table_length, seq_table_name);
   if (!seqTable->LoadVGMFile()) {
     delete seqTable;
     seqTable = NULL;
@@ -217,164 +237,4 @@ void CPSScanner::Scan(RawFile *file, void *info) {
   }
 
   return;
-}
-
-void CPSScanner::LoadCPS1(MAMEGame *gameentry) {
-
-  CPSFormatVer fmt_ver = GetVersionEnum(gameentry->fmt_version_str);
-
-  CPS1SampleInstrSet *instrset = nullptr;
-  CPS1SampColl *sampcoll = nullptr;
-
-  MAMERomGroup* seqRomGroupEntry = gameentry->GetRomGroupOfType("audiocpu");
-  if (!seqRomGroupEntry)
-    return;
-  uint32_t seq_table_offset;
-  uint32_t seq_table_length;
-  if (!seqRomGroupEntry->file ||
-      !seqRomGroupEntry->GetHexAttribute("seq_table", &seq_table_offset))
-    return;
-
-  if (fmt_ver == VER_UNDEFINED) {
-    string alert = "XML entry uses an undefined QSound version: " + gameentry->fmt_version_str;
-    pRoot->AddLogItem(new LogItem(alert, LOG_LEVEL_ERR, "CPSScanner"));
-    return;
-  }
-
-  RawFile *programFile = seqRomGroupEntry->file;
-
-  MAMERomGroup* sampsRomGroupEntry = gameentry->GetRomGroupOfType("oki6295");
-  if (sampsRomGroupEntry && sampsRomGroupEntry->file) {
-    uint32_t instrTablePtrOffset = 4;
-    uint32_t instr_table_offset = programFile->GetShortBE(seq_table_offset + instrTablePtrOffset);
-
-    RawFile *samplesFile = sampsRomGroupEntry->file;
-
-    ostringstream name;
-    name.str("");
-    name << gameentry->name.c_str() << " oki msm6295 instrument set";
-    auto instrset_name = name.str();
-    name.str("");
-    name << gameentry->name.c_str() << " sample collection";
-    auto sampcoll_name = name.str();
-
-    instrset = new CPS1SampleInstrSet(programFile,
-                               fmt_ver,
-                               instr_table_offset,
-                               instrset_name);
-    if (!instrset->LoadVGMFile()) {
-      delete instrset;
-      instrset = NULL;
-    }
-
-    sampcoll = new CPS1SampColl(samplesFile, instrset, 0, samplesFile->size(), sampcoll_name);
-    if (!sampcoll->LoadVGMFile()) {
-      delete sampcoll;
-      sampcoll = NULL;
-    }
-  }
-
-  string seq_table_name;
-  ostringstream name;
-
-  name.str("");
-  name << gameentry->name.c_str() << " sequence pointer table";
-  seq_table_name = name.str();
-
-  uint8_t ptrsStart;
-  const uint8_t ptrSize = 2;
-
-  switch (fmt_ver) {
-    case VER_CPS1_500: ptrsStart = 2; break;
-    case VER_CPS1_200ff: ptrsStart = 3; break;
-    default: ptrsStart = 4;
-  }
-
-  switch (fmt_ver) {
-    case VER_CPS1_200:
-    case VER_CPS1_200ff:
-    case VER_CPS1_350:
-    case VER_CPS1_425:
-    case VER_CPS1_500:
-      seq_table_length = (uint32_t) (programFile->GetByte(seq_table_offset) * 2) + ptrsStart;
-      break;
-    case VER_CPS1_502:
-      seq_table_length = (uint32_t) (programFile->GetShortBE(seq_table_offset) * 2) + ptrsStart;
-      break;
-  }
-
-  unsigned int k = 0;
-  uint32_t seqPointer = 0;
-
-  // Add SeqTable as Miscfile
-  VGMMiscFile *seqTable = new VGMMiscFile(CPSFormat::name, seqRomGroupEntry->file, seq_table_offset, seq_table_length, seq_table_name);
-  if (!seqTable->LoadVGMFile()) {
-    delete seqTable;
-    seqTable = NULL;
-  }
-
-  int seqNum = 0;
-  for (k = ptrsStart; (seq_table_length == 0 || k < seq_table_length); k += ptrSize) {
-
-    seqPointer = programFile->GetShortBE(seq_table_offset + k);
-
-    if (seqPointer == 0) {
-      continue;
-    }
-
-    seqTable->AddSimpleItem(seq_table_offset + k, ptrSize, "Sequence Pointer");
-
-    name.str("");
-    name << gameentry->name.c_str() << " seq " << seqNum;
-    string seqName = name.str();
-    CPSSeq *newSeq = new CPSSeq(programFile, seqPointer, fmt_ver, seqName);
-
-    if (!newSeq->LoadVGMFile()) {
-      delete newSeq;
-      continue;
-    }
-
-    name.str("");
-    name << gameentry->name.c_str() << " song " << seqNum++;
-    VGMColl* coll = new VGMColl(name.str());
-
-    coll->UseSeq(newSeq);
-    coll->AddInstrSet(instrset);
-    coll->AddSampColl(sampcoll);
-    if (!coll->Load()) {
-      delete coll;
-    }
-  }
-}
-
-CPSFormatVer CPSScanner::GetVersionEnum(string &versionStr) {
-  if (versionStr == "CPS1_2.00") return VER_CPS1_200;
-  if (versionStr == "CPS1_2.00ff") return VER_CPS1_200ff;
-  if (versionStr == "CPS1_3.50") return VER_CPS1_350;
-  if (versionStr == "CPS1_4.25") return VER_CPS1_425;
-  if (versionStr == "CPS1_5.00") return VER_CPS1_500;
-  if (versionStr == "CPS1_5.02") return VER_CPS1_502;
-  if (versionStr == "1.00") return VER_100;
-  if (versionStr == "1.01") return VER_101;
-  if (versionStr == "1.03") return VER_103;
-  if (versionStr == "1.04") return VER_104;
-  if (versionStr == "1.05a") return VER_105A;
-  if (versionStr == "1.05c") return VER_105C;
-  if (versionStr == "1.05") return VER_105;
-  if (versionStr == "1.06b") return VER_106B;
-  if (versionStr == "1.15c") return VER_115C;
-  if (versionStr == "1.15") return VER_115;
-  if (versionStr == "1.16b") return VER_116B;
-  if (versionStr == "1.16") return VER_116;
-  if (versionStr == "1.30") return VER_130;
-  if (versionStr == "1.31") return VER_131;
-  if (versionStr == "1.40") return VER_140;
-  if (versionStr == "1.71") return VER_171;
-  if (versionStr == "1.80") return VER_180;
-  if (versionStr == "2.00") return VER_200;
-  if (versionStr == "2.01b") return VER_201B;
-  if (versionStr == "2.10") return VER_210;
-  if (versionStr == "2.11") return VER_211;
-  if (versionStr == "CPS3") return VER_CPS3;
-  return VER_UNDEFINED;
 }
